# Section authorization: Optimize method to make it faster